pyproject.toml
[project]
name = "pong"
version = "0.1.0"
description = "Pong RL"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "ale-py>=0.11.2",
    "array-api-compat>=1.12.0",
    "autorom[accept-rom-license]>=0.6.1",
    "black>=25.11.0",
    "gymnasium[accept-rom-license,atari,other]>=1.2.2",
    "hydra-core>=1.3.2",
    "imageio>=2.37.2",
    "matplotlib>=3.10.7",
    "moviepy==1.0.3",
    "mypy>=1.18.2",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "ruff>=0.14.3",
    "seaborn>=0.13.2",
    "shimmy>=2.0.0",
    "tensorboard>=2.20.0",
    "tensorboard-plugin-profile>=2.21.1",
    "torch>=2.9.0",
    "torchinfo>=1.8.0",
    "torchviz>=0.0.3",
    "tqdm>=4.67.1",
    "types-seaborn>=0.13.2.20250914",
    "types-tqdm>=4.67.0.20250809",
]
